server/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psutil==5.9.8
python-oracledb==2.2.1
aiosqlite==0.20.0
pydantic-settings==2.2.1
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
